Port Rule to gsuite_activityevents
