+++ gsuite_activityevent_rules/gsuite_brute_force_login.yml
@@ -1,0 +1,58 @@
+AnalysisType: rule
+Filename: gsuite_brute_force_login.py
+RuleID: GSuite.BruteForceLogin
+DisplayName: --DEPRECATED-- GSuite Brute Force Login
+Enabled: false
+LogTypes:
+  - GSuite.ActivityEvent
+Tags:
+  - GSuite
+Severity: Medium
+Threshold: 10
+DedupPeriodMinutes: 10
+Description: A GSuite user was denied login access several times
+Reference: https://developers.google.com/admin-sdk/reports/v1/appendix/activity/login#login_failure
+Runbook: Analyze the IP they came from and actions taken before/after.
+Tests:
+  -
+    Name: Failed Login
+    ExpectedResult: true
+    Log:
+      {
+        "id": {
+          "applicationName": "login",
+        },
+        "actor": {
+          "email": "some.user@somedomain.com",
+        },
+        "type": "login",
+        "name": "login_failure",
+      }
+  -
+    Name: Successfull Login
+    ExpectedResult: false
+    Log:
+      {
+        "id": {
+          "applicationName": "login",
+        },
+        "actor": {
+          "email": "some.user@somedomain.com",
+        },
+        "type": "login",
+        "name": "login_success",
+      }
+  -
+    Name: Other Login Event
+    ExpectedResult: false
+    Log:
+      {
+        "id": {
+          "applicationName": "login",
+        },
+        "actor": {
+          "email": "some.user@somedomain.com",
+        },
+        "type": "login",
+        "name": "login_verification",
+      }
