--- gsuite_reports_rules/gsuite_brute_force_login.yml
@@ -1,84 +1,0 @@
-AnalysisType: rule
-Filename: gsuite_brute_force_login.py
-RuleID: GSuite.BruteForceLogin
-DisplayName: --DEPRECATED-- GSuite Brute Force Login
-Enabled: false
-LogTypes:
-  - GSuite.Reports
-Tags:
-  - GSuite
-Severity: Medium
-Threshold: 10
-DedupPeriodMinutes: 10
-Description: A GSuite user was denied login access several times
-Reference: https://developers.google.com/admin-sdk/reports/v1/appendix/activity/login#login_failure
-Runbook: Analyze the IP they came from and actions taken before/after.
-Tests:
-  -
-    Name: Failed Login
-    ExpectedResult: true
-    Log:
-      {
-        "id": {
-          "applicationName": "login",
-          "customerId": "E0391azke",
-          "time": "2021-01-03 04:23:26.530000000",
-          "uniqueQualifier": "229944411100"
-        },
-        "actor": {
-          "email": "nick@acme.io",
-          "profileId": "11949494222400014922"
-        },
-        "kind": "admin#reports#activity",
-        "ipAddress": "2604:2d80:490e:e400:d1f6:90dd:56af:bb42",
-        "events": [
-          {
-            "name": "login_failure",
-            "parameters": [
-              {
-                "name": "login_type",
-                "value": "google_password"
-              },
-              {
-                "multiValue": [
-                  "password"
-                ],
-                "name": "login_challenge_method"
-              }
-            ],
-            "type": "login"
-          }
-        ],
-        "p_event_time": "2021-01-03 04:23:26.530",
-        "p_parse_time": "2021-01-03 05:46:50.364",
-        "p_log_type": "GSuite.Reports",
-        "p_row_id": "5e3d433a78dqc2a2c3f8f59d07a49201",
-        "p_any_ip_addresses": [
-          "2604:2d80:490e:e400:d1f6:90dd:56af:bb42"
-        ]
-      }
-  -
-    Name: Normal Login Event
-    ExpectedResult: false
-    Log:
-      {
-        'id': {'applicationName': 'login'},
-        'events': [
-          {
-            'type': 'login',
-            'name': 'login_success'
-          }
-        ]
-      }
-  -
-    Name: Non Login Event
-    ExpectedResult: false
-    Log:
-      {
-        'id': {'applicationName': 'login'},
-        'events': [
-          {
-            'type': 'account_warning',
-          }
-        ]
-      }
